Fix launch-linux.sh / launch-linux-dev.sh file warnings
Before deleting a specific file/directory, that file/directory's existence is now checked.

Printed warnings are now written to STDERR instead of STDOUT.

Git operations are now skipped when SwarmUI is installed incorrectly, as these would fail anyway.

+++ launch-linux-dev.sh
@@ -11,16 +11,16 @@ export DOTNET_CLI_TELEMETRY_OPTOUT=1
 export DOTNET_SYSTEM_GLOBALIZATION_INVARIANT=1
 
 # Cycle build folder forward
-rm -rf ./src/bin/live_release_backup
-mv ./src/bin/live_release ./src/bin/live_release_backup
-rm ./src/bin/must_rebuild
-rm ./src/bin/last_build
+[ -e ./src/bin/live_release_backup ] && rm -rf ./src/bin/live_release_backup
+[ -e ./src/bin/live_release ] && mv ./src/bin/live_release ./src/bin/live_release_backup
+[ -e ./src/bin/must_rebuild ] && rm ./src/bin/must_rebuild
+[ -e ./src/bin/last_build ] && rm ./src/bin/last_build
 
 # Build the program
-dotnet build src/SwarmUI.csproj --configuration Debug -o ./src/bin/live_release
+dotnet build ./src/SwarmUI.csproj --configuration Debug -o ./src/bin/live_release
 
 # Default env configuration, gets overwritten by the C# code's settings handler
 export ASPNETCORE_ENVIRONMENT="Production"
 export ASPNETCORE_URLS="http://*:7801"
 # Actual runner.
-dotnet src/bin/live_release/SwarmUI.dll $@
+dotnet ./src/bin/live_release/SwarmUI.dll $@
